email_reader/requirements.txt
boto3
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
